refactor(search): migrate 7 sites off core.search shim and delete it (#66)
Removes the core/search/__init__.py compatibility shim and migrates all
callers to import directly from core.services.search_service.

Touch points (7 edits + 1 deletion):
- core/db/crud.py: 4 function-local imports (search_memory_blocks_enhanced,
  _fulltext, _semantic, _hybrid)
- core/db/repositories/memory_blocks.py: 1 module-top import
- core/search/evaluation.py: 1 module-top import (kept in place; the
  module survives in a PEP-420 namespace package)
- tests/unit/test_crud_query_expansion.py: monkeypatch target updated
  in lockstep with crud.py — without this the patch raises AttributeError
  at setup once the shim is gone (verified by subagent design-experimenter
  empirical pass)
- core/search/__init__.py: deleted (-9 LOC)

The original issue listed stale crud.py line numbers (615/738/785/833 →
actually 645/768/815/863) and missed the test monkeypatch as a 7th touch
point. Caught by playbook verification step before execution.

core/search/ becomes an implicit namespace package (PEP 420) containing
only evaluation.py, which has 3 external importers (a script, an
integration test, two README mentions) and stays put. SearchService (the
class) was already imported directly from core.services.search_service
everywhere, so only get_search_service flowed through the shim.

Net diff: 5 files changed, +7/-17 (net -10 LOC).

Verification (run locally; integration branch is CI-silent until staging
PR per .github/workflows/deploy.yml):
- uv run pytest tests/unit/test_crud_query_expansion.py -v: 2 passed
  (FakeService confirmed exercised via calls assertion)
- uv run pytest (full suite): 840 passed, 8 skipped, 81.06% coverage
- python -c "from core.search.evaluation import evaluate_cases": ok
  (namespace package resolves)
- Audit greps (static + identifier + dynamic + monkeypatch/patch
  targets): only core.search.evaluation submodule hits remain

Refs #66, umbrella #85.
