--- apps/hindsight-service/core/search/__init__.py
@@ -1,10 +1,0 @@
-"""
-Search module compatibility shim.
-
-New code should import from `core.services.search_service`.
-This module re-exports the service for backward compatibility.
-"""
-
-from core.services.search_service import SearchService, get_search_service  # type: ignore
-
-__all__ = ["SearchService", "get_search_service"]
